Addign post and delete post using ajax

+++ controllers/posts_controller.js
@@ -1,7 +1,8 @@
 const { localsName } = require('ejs');
 const { findById } = require('../models/postSchema');
 const Post = require('../models/postSchema');
-const Comment = require('../models/commetSchema')
+const Comment = require('../models/commetSchema');
+const User = require('../models/userSchema');
 
 
 // Funtion to upload new post
@@ -16,7 +17,16 @@ module.exports.create = async function (req, res) {
                 user: req.user._id,
                 name: req.body.name
             });
-
+            let UserDetails = await User.findById(post.user);
+        if (req.xhr) {
+            return res.status(200).json({
+                data: {
+                    post: post,
+                    UserDetails : UserDetails
+                },
+                message: "Post Created!"
+            })
+        }
         req.flash('success', 'Post Added Successfully')
         return res.redirect('back')
     }
@@ -36,13 +46,21 @@ module.exports.destroy = async function (req, res) {
             let comments = await Comment.findById(comment._id);
             comments.remove();
         }
-
+        if(req.xhr)
+        {
+            return res.status(200).json({
+                data : {
+                    post_id : req.params.id
+                },
+                message : 'Post Deleted'
+            })
+        }
         // .id means converting the object id into string inbuilt  
         if (post.user == req.user.id) {
 
             post.remove();
             req.flash('success', 'Post Deleted Successfully')
-            Comment.deleteMany({ post: req.params.id }, function (err) {
+            await Comment.deleteMany({ post: req.params.id }, function (err) {
                 return res.redirect('back');
             })
